Calculate money rate has been edited

+++ app/src/main/java/com/example/madcamp_week2/CrewCardActivity.kt
@@ -195,7 +195,7 @@ fun CalculateMoney(crewData: List<String>) : List<Float> {
     var UsedMoney = crewData[crewData.size - 1].toInt()
 
     var RemainMoney = TarMoney - UsedMoney
-    var RemainMoneyRate = (UsedMoney/TarMoney).toFloat()
+    var RemainMoneyRate = UsedMoney.toFloat()/TarMoney.toFloat()
 
     return listOf(RemainMoney.toFloat(), RemainMoneyRate)
 }
